[navigation] dynamic routing
* [wip] start adding dynamic routing to case and eligibility screens.

* [feat] implement dynamic routing for the CaseSummaryScreen.

* [feat] implement dynamic routing on Forms screens.

* [feat] implement dynamic routing in AddCase flow.

* [cleanup] finishing touches.
